Fix bug in mrDMD reconstruction due to null eigenvalues
When choosing a large number of resolution levels, the upper bins could not fit in the tree structure. In order to handle this situation, the previous implementation forced the DMD modes, amplitudes, eigenvalues, Atilda, etc. of such upper (non fitting) bins to be zero. This was done through a `try...except` statement. However, this solution was causing a problem of division by zero during the reconstruction of the dynamics. In the present commit, a threshold for the maximum resolution level was stabilished, thus eliminating the division by zero issue.