Add exception for MemoryError in mrDMD reconstruction
When the size of the snapshot matrix is big and the number of resolution levels in mrDMD is large, a MemoryError may occur.
This happens because all partial reconstructed data are stored in the buffer before doing the summation. Although being fast, this procedure can present memory limitations.
In order to circumvent this problem, an exception for MemoryError was implemented, in which the partial data is successively added up through an explicit loop.