Major upgrade to the conjugate pair eig constraint
Previous eigenvalue constraint for conjugate pairs would raise errors
when trying to fit systems which should be conjugate pairs. The reason
for this is that the solver can "wander" off into states for which
the conjugate pairs can be more ambiguous than what the original
constraint function considered. Many of these problematic states are
now considered in a separate test for the conjugate pair constrainer.

The new constraint matches the closest conjugate pairs, removes
them from consideration, and then finds the next closest. It is also
has handling for a wide range of edge cases found from fitting
thousands of DMD models as a result of mrCOSTS.